+++ lib/minitest-activemodel/matchers/activemodel/validate_length_matcher.rb
@@ -7,12 +7,15 @@ module MiniTest
       # * <tt>with_minimum</tt> - minimum length of the attribute.
       #   Aliased as: <tt>with_min</tt> and <tt>is_at_least</tt>.
       # * <tt>with_maximum</tt> - maximum length of the attribute.
+      #   Aliased as: <tt>with_max</tt> and <tt>is_at_most</tt>.
       #
       #   it { must validate_length_of :name }
       #   it { must validate_length_of(:name).with_minimum(10) }
       #   it { must validate_length_of(:name).with_min(10) }
       #   it { must validate_length_of(:name).is_at_least(10) }
       #   it { must validate_length_of(:name).with_maximum(100) }
+      #   it { must validate_length_of(:name).with_max(100) }
+      #   it { must validate_length_of(:name).is_at_most(100) }
       def validate_length_of attr
         ValidateLengthMatcher.new attr
       end
@@ -36,6 +39,8 @@ module MiniTest
           @maximum = value
           self
         end
+        alias :with_max   :with_maximum
+        alias :is_at_most :with_maximum
 
         def matches? subject
           return false unless @result = super(subject)